Todos! More progress..

+++ static/js/develop/controllers/feedcontroller.js
@@ -3,13 +3,8 @@
 (function() {
   angular.module("projecto").controller(
     "FeedController", ["$scope", "FeedService", "ProjectsService", function($scope, FeedService, ProjectsService) {
-      $scope.currentProject = null;
       $scope.posts = [];
       $scope.newpost = "";
-
-      var notLoaded = function() {
-        $("body").statusmsg("open", "The page has not been fully loaded yet. Please wait...", {type: "warning", closable: true});
-      };
 
       $scope.post = function() {
         if ($scope.newpost && $scope.currentProject) {
@@ -35,18 +30,20 @@
 
       $scope.deletePost = function(post) {
         if ($scope.currentProject) {
-          FeedService.delete($scope.currentProject, post).done(function(){
-            for (var i=0; i<$scope.posts.length; i++) {
-              if ($scope.posts[i].key === post.key) {
-                $scope.$apply(function(){
-                  $scope.posts.splice(i, 1);
-                });
-                break;
+          if (confirm("Are you sure you want to delete this post?")) {
+            FeedService.delete($scope.currentProject, post).done(function(){
+              for (var i=0; i<$scope.posts.length; i++) {
+                if ($scope.posts[i].key === post.key) {
+                  $scope.$apply(function(){
+                    $scope.posts.splice(i, 1);
+                  });
+                  break;
+                }
               }
-            }
-          }).fail(function(xhr){
-            $("body").statusmsg("open", "Delete error " + xhr.status, {type: "error", closable: true});
-          });
+            }).fail(function(xhr){
+              $("body").statusmsg("open", "Delete error " + xhr.status, {type: "error", closable: true});
+            });
+          }
         } else {
           notLoaded();
         }
@@ -66,9 +63,11 @@
         }
       };
 
+      $scope.currentProject = null;
+
       ProjectsService.getCurrentProject().done(function(currentProject){
         $scope.currentProject = currentProject;
-        $scope.update(currentProject);
+        $scope.update();
         $scope.$$phase || $scope.$apply();
       });
     }]
